+++ ecommerce-engrenagem/src/main/java/application/entities/BaseEntity.java
@@ -1,0 +1,25 @@
+package application.entities;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import java.io.Serializable;
+
+@Getter
+@Setter
+public class BaseEntity implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    @Id
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
+    private Long id;
+
+    public void validate(){
+        if(id == null){
+            throw new IllegalArgumentException("Name cannot be null");
+        }
+    }
+}
